+++ lib/vector_number/special_unit.rb
@@ -1,18 +1,19 @@
 # frozen_string_literal: true
 
 class VectorNumber
-  # Class for representing special units.
+  # Class for representing special (unique) units.
   #
-  # The public API consists of:
-  # - +#==+/+#eql?+/+#equal?+ (from +Object+)
-  # - +#hash+ (from +Object+)
-  # - {#to_s}
-  # - {#inspect}
-  # - {#pretty_print}
+  # There usually isn't much point in using these, aside from {VectorNumber::NUMERIC_UNITS}.
+  # However, they can be helpful to denote units that aren't equal to any other object.
+  #
+  # +VectorNumber#to_s+ (and +inspect+) will use {#to_s} text instead of {#inspect} text
+  # for these units by default, without a multiplication operator. However, consider
+  # using customization of +VectorNumber#to_s+ instead if this is what you want.
   #
   # @since 0.6.0
   class SpecialUnit
-    # @api private
+    # Returns a new, unique unit, not equal to any other unit.
+    #
     # @param unit [#to_s] name for {#inspect}
     # @param text [String] text for {#to_s}
     def initialize(unit, text = unit.to_s)